# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awards Management API",
        "description": "API for managing awards, nominees, voting, and events",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Awards Management API Server"
        }
    ],
    "paths": {
...
        "/api/v1/nominees": {
            "get": {
                "tags": [
                    "Nominees"
                ],
                "summary": "Get list of nominees",
                "description": "Returns list of nominees with filtering options",
                "operationId": "getNominees",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "winners_only",
                        "in": "query",
                        "description": "Show only winners",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Order by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "votes",
                            "enum": [
                                "votes",
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order_direction",
                        "in": "query",
                        "description": "Order direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "organization": {
                                                        "type": "string",
                                                        "example": "Tech Corp"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Best Innovation Award"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "votes": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "voting_percentage": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15.5
                                                    },
                                                    "can_vote": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "is_winner": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "impact_summary": {
                                                        "type": "string"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "New York, USA"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}